requirements.txt
python-etcd>=0.4.3
lxml!=3.7.0,>=2.3
# libvirt-python
